Add mock for playbook so dagger can create the worker factory map.
